pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dara"
version = "0.1.0"
description = "Detect & Assist Recognition AI"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch",
    "transformers",
    "pillow",
    "gtts",
    "deep-translator",
    "huggingface_hub",
    "scipy",
    "einops",
    "timm",
    "peft"
]

[project.urls]
"Homepage" = "https://github.com/ardelyo/dara"
"Bug Tracker" = "https://github.com/ardelyo/dara/issues"

[tool.setuptools.packages.find]
where = ["src"]
